kanban_backend/requirements.txt
aniso8601==9.0.1
Babel==2.11.0
bcrypt==4.0.1
blinker
click
dataclasses
dnspython
email-validator
bleach
Flask-Cors
Flask
Flask-BabelEx
Flask-Login
Flask-Mail
Flask-Principal
Flask-RESTful
Flask-Security-Too
Flask-SQLAlchemy
Flask-WTF
greenlet
idna
importlib-metadata
itsdangerous
Jinja2
MarkupSafe
passlib
pkg_resources
pytz
six
speaklater
SQLAlchemy
typing_extensions
Werkzeug
WTForms
zipp
redis
celery[redis]
flask-sse